PR #13897 did a great job at improving the performance of running `work_packages/update_ancestors_service_spec` spec by making it 70% faster.

With some additional `shared_let`, it's possible to have less work packages created and gain a little 20% more.

Run with `FPROF=1` to have below statistics.

Before:

```
Finished in 4.55 seconds (files took 8.14 seconds to load)
65 examples, 0 failures

[TEST PROF INFO] Time spent in factories: 00:02.833 (39.19% of total time)
[TEST PROF INFO] Factories usage

 Total: 254
 Total top-level: 227
 Total time: 00:02.833 (out of 00:10.861)
 Total uniq factories: 7

   total   top-level     total time      time per call      top-level time               name

     221         221        2.6216s            0.0119s             2.6216s       work_package
       7           2        0.1100s            0.0157s             0.0777s               user
       7           0        0.0167s            0.0024s             0.0000s notification_setting
       6           1        0.2549s            0.0425s             0.1045s project_with_types
       6           1        0.0245s            0.0041s             0.0130s           priority
       6           1        0.0244s            0.0041s             0.0142s             status
       1           1        0.0024s            0.0024s             0.0024s      closed_status
```

After:

```
Finished in 3.59 seconds (files took 8.25 seconds to load)
65 examples, 0 failures

[TEST PROF INFO] Time spent in factories: 00:01.857 (29.59% of total time)
[TEST PROF INFO] Factories usage

 Total: 131
 Total top-level: 129
 Total time: 00:01.857 (out of 00:09.947)
 Total uniq factories: 7

   total   top-level     total time      time per call      top-level time               name

     123         123        1.6649s            0.0135s             1.6649s       work_package
       2           2        0.0780s            0.0390s             0.0780s               user
       2           0        0.0112s            0.0056s             0.0000s notification_setting
       1           1        0.0946s            0.0946s             0.0946s project_with_types
       1           1        0.0084s            0.0084s             0.0084s           priority
       1           1        0.0096s            0.0096s             0.0096s             status
       1           1        0.0017s            0.0017s             0.0017s      closed_status
```